# Gateway/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Jolt-API",
    "description": "Documentation de l’API Jolt (microservices)",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token d'authentification (Bearer <token>) ou dans les cookies"
    }
  },
  "paths": {
    "/uploads/maintains/{filename}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}